chore: move FocusZone to react-bindings
