+++ packages/react-bindings/src/index.ts
@@ -1,5 +1,14 @@
-export { default as getElementType } from './utils/getElementType'
-export { default as getUnhandledProps } from './utils/getUnhandledProps'
-
 export { default as unstable_useDispatchEffect } from './hooks/useDispatchEffect'
 export { default as useStateManager } from './hooks/useStateManager'
+
+export { default as AutoFocusZone } from './FocusZone/AutoFocusZone'
+export * from './FocusZone/AutoFocusZone.types'
+export { default as FocusTrapZone } from './FocusZone/FocusTrapZone'
+export * from './FocusZone/FocusTrapZone.types'
+export { default as FocusZone } from './FocusZone/FocusZone'
+export * from './FocusZone/FocusZone.types'
+export * from './FocusZone/focusUtilities'
+
+export { default as callable } from './utils/callable'
+export { default as getElementType } from './utils/getElementType'
+export { default as getUnhandledProps } from './utils/getUnhandledProps'
