+++ packages/react-bindings/src/utils/callable.ts
@@ -1,4 +1,3 @@
-// https://jsperf.com/startdust-callable
 const callable = (possibleFunction: any) => (...args: any[]) => {
   return typeof possibleFunction === 'function' ? possibleFunction(...args) : possibleFunction
 }
