+++ packages/react/src/components/Portal/Portal.tsx
@@ -1,4 +1,5 @@
 import { AccessibilityAttributes } from '@stardust-ui/accessibility'
+import { FocusTrapZone, FocusTrapZoneProps } from '@stardust-ui/react-bindings'
 import { EventListener } from '@stardust-ui/react-component-event-listener'
 import { handleRef, Ref, toRefObject } from '@stardust-ui/react-component-ref'
 import * as customPropTypes from '@stardust-ui/react-proptypes'
@@ -16,7 +17,6 @@ import {
   rtlTextContainer,
 } from '../../lib'
 import PortalInner from './PortalInner'
-import { FocusTrapZone, FocusTrapZoneProps } from '../../lib/accessibility/FocusZone'
 import { AccessibilityHandlerProps } from '../../lib/accessibility/reactTypes'
 
 export type TriggerAccessibility = {
