+++ packages/react/src/components/Tree/TreeItem.tsx
@@ -1,3 +1,4 @@
+import { Accessibility, treeItemBehavior } from '@stardust-ui/accessibility'
 import * as customPropTypes from '@stardust-ui/react-proptypes'
 import * as _ from 'lodash'
 import * as PropTypes from 'prop-types'
@@ -5,7 +6,6 @@ import * as React from 'react'
 
 import { Ref } from '@stardust-ui/react-component-ref'
 import TreeTitle, { TreeTitleProps } from './TreeTitle'
-import { Accessibility, treeItemBehavior } from '@stardust-ui/accessibility'
 import {
   UIComponent,
   childrenExist,
