--- packages/react/src/lib/accessibility/FocusZone/index.ts
@@ -1,7 +1,0 @@
-export { default as FocusZone } from './FocusZone'
-export * from './FocusZone.types'
-export { default as FocusTrapZone } from './FocusTrapZone'
-export * from './FocusTrapZone.types'
-export * from './focusUtilities'
-export { default as AutoFocusZone } from './AutoFocusZone'
-export * from './AutoFocusZone.types'
